requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --build-isolation --output-file=requirements.txt requirements.in
#
altair==4.1.0             # via -r requirements.in, streamlit
appdirs==1.4.3            # via black
astor==0.8.1              # via streamlit
astunparse==1.6.3         # via pyastsim
attrs==19.3.0             # via black, jsonschema
backcall==0.1.0           # via ipython
base58==2.0.0             # via streamlit
black==19.10b0            # via -r requirements.in, nb-black
bleach==3.1.5             # via nbconvert
blinker==1.4              # via streamlit
boto3==1.13.1             # via streamlit
botocore==1.16.1          # via boto3, s3transfer, streamlit
cachetools==4.1.0         # via streamlit
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via black, pip-tools, streamlit
cycler==0.10.0            # via matplotlib
cython==0.29.17           # via -r requirements.in
decorator==4.4.2          # via ipython, traitlets, validators
defusedxml==0.6.0         # via nbconvert
docutils==0.15.2          # via botocore
editdistance==0.5.3       # via pyastsim
entrypoints==0.3          # via altair, flake8, nbconvert
enum-compat==0.0.3        # via streamlit
flake8==3.7.9             # via -r requirements.in
idna==2.9                 # via requests
importlib-metadata==1.6.0  # via jsonschema
ipykernel==5.2.1          # via -r requirements.in, ipywidgets, jupyter, jupyter-console, notebook, pydeck, qtconsole
ipython-genutils==0.2.0   # via jupyter-contrib-nbextensions, nbformat, notebook, qtconsole, traitlets
ipython==7.14.0           # via -r requirements.in, ipykernel, ipywidgets, jupyter-console, jupyter-latex-envs, nb-black
ipywidgets==7.5.1         # via jupyter, pydeck
jedi==0.17.0              # via ipython
jinja2==2.11.2            # via altair, nbconvert, notebook, pydeck
jmespath==0.9.5           # via boto3, botocore
joblib==0.14.1            # via scikit-learn
jsonschema==3.2.0         # via altair, nbformat
jupyter-client==6.1.3     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==6.1.0    # via jupyter
jupyter-contrib-core==0.3.3  # via jupyter-contrib-nbextensions, jupyter-nbextensions-configurator
jupyter-contrib-nbextensions==0.5.1  # via -r requirements.in
jupyter-core==4.6.3       # via jupyter-client, jupyter-contrib-core, jupyter-contrib-nbextensions, jupyter-latex-envs, jupyter-nbextensions-configurator, nbconvert, nbformat, notebook, qtconsole
jupyter-highlight-selected-word==0.2.0  # via jupyter-contrib-nbextensions
jupyter-latex-envs==1.4.6  # via jupyter-contrib-nbextensions
jupyter-nbextensions-configurator==0.4.1  # via -r requirements.in, jupyter-contrib-nbextensions
jupyter==1.0.0            # via -r requirements.in
kiwisolver==1.2.0         # via matplotlib
llvmlite==0.32.0          # via numba
lxml==4.5.0               # via jupyter-contrib-nbextensions
markupsafe==1.1.1         # via jinja2
matplotlib==3.2.1         # via seaborn
mccabe==0.6.1             # via flake8
mistune==0.8.4            # via nbconvert
mypy-extensions==0.4.3    # via mypy
mypy==0.770               # via -r requirements.in
nb-black==1.0.7           # via -r requirements.in
nbconvert==5.6.1          # via jupyter, jupyter-contrib-nbextensions, jupyter-latex-envs, notebook
nbformat==5.0.6           # via ipywidgets, nbconvert, notebook
notebook==6.0.3           # via jupyter, jupyter-contrib-core, jupyter-contrib-nbextensions, jupyter-latex-envs, jupyter-nbextensions-configurator, widgetsnbextension
numba==0.49.0             # via umap-learn
numpy==1.18.4             # via -r requirements.in, altair, matplotlib, numba, pandas, pydeck, scikit-learn, scipy, seaborn, streamlit, umap-learn
packaging==20.3           # via bleach, streamlit
pandas==1.0.3             # via -r requirements.in, altair, seaborn, streamlit
pandocfilters==1.4.2      # via nbconvert
parso==0.7.0              # via jedi
pathspec==0.8.0           # via black
pathtools==0.1.2          # via watchdog
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.1.2             # via streamlit
pip-tools==5.1.1          # via -r requirements.in
prometheus-client==0.7.1  # via notebook
prompt-toolkit==3.0.5     # via ipython, jupyter-console
protobuf==3.11.3          # via streamlit
ptyprocess==0.6.0         # via pexpect, terminado
pyastsim==1.1.0           # via -r requirements.in
pycode-similar==1.3       # via -r requirements.in
pycodestyle==2.5.0        # via flake8
pydeck==0.3.1             # via streamlit
pyflakes==2.1.1           # via flake8
pygments==2.6.1           # via ipython, jupyter-console, nbconvert, qtconsole
pyparsing==2.4.7          # via matplotlib, packaging
pyrsistent==0.16.0        # via jsonschema
python-dateutil==2.8.1    # via botocore, jupyter-client, matplotlib, pandas, streamlit
pytz==2020.1              # via pandas, tzlocal
pyyaml==5.3.1             # via jupyter-contrib-nbextensions, jupyter-nbextensions-configurator
pyzmq==19.0.0             # via jupyter-client, notebook, qtconsole
qtconsole==4.7.3          # via jupyter
qtpy==1.9.0               # via qtconsole
regex==2020.4.4           # via black
requests==2.23.0          # via streamlit
s3transfer==0.3.3         # via boto3
scikit-learn==0.22.2.post1  # via -r requirements.in, umap-learn
scipy==1.4.1              # via scikit-learn, seaborn, umap-learn
seaborn==0.10.1           # via -r requirements.in
send2trash==1.5.0         # via notebook
six==1.14.0               # via astunparse, bleach, cycler, jsonschema, packaging, pip-tools, protobuf, pyrsistent, python-dateutil, traitlets, validators
streamlit==0.58.0         # via -r requirements.in
tbb==2020.0.133           # via umap-learn
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
toml==0.10.0              # via black, streamlit
toolz==0.10.0             # via altair
tornado==5.1.1            # via ipykernel, jupyter-client, jupyter-contrib-core, jupyter-contrib-nbextensions, jupyter-nbextensions-configurator, notebook, streamlit, terminado
tqdm==4.46.0              # via -r requirements.in
traitlets==4.3.3          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-contrib-core, jupyter-contrib-nbextensions, jupyter-core, jupyter-latex-envs, jupyter-nbextensions-configurator, nbconvert, nbformat, notebook, pydeck, qtconsole
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.2  # via mypy
tzlocal==2.0.0            # via streamlit
umap-learn==0.4.2         # via -r requirements.in
urllib3==1.25.9           # via botocore, requests
validators==0.14.3        # via streamlit
watchdog==0.10.2          # via streamlit
wcwidth==0.1.9            # via prompt-toolkit
webencodings==0.5.1       # via bleach
wheel==0.34.2             # via astunparse
widgetsnbextension==3.5.1  # via ipywidgets
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
